refactor: publications container moved to components for better organisation
