+++ frontend/src/components/PublicationsContainer.jsx
